requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pyqtgraph==0.13.7
pyserial==3.5
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
shiboken6==6.7.0
paho-mqtt==2.1.0